requirements.txt
django
gunicorn
django-heroku
requests
psycopg2-binary
dj-database-url
Pillow